refactor(test): migrate create-run.test.ts to route integration test (#9415)
Migrate 24 tests from service-level create-run.test.ts into the existing
route.test.ts for POST /api/agent/runs. Tests now call the route handler
instead of internal createCliRun() service function.

- 24 tests migrated (happy path, concurrency, memory, artifacts, volumes,
  runner routing, experimental profiles, org resolution)
- 14 tests already covered by existing route.test.ts
- 2 exceptions documented (advisory lock serialization, CLI-only callbacks)
- Delete source test file and remove unused createCliRun/CliRunParams
